refactor: moved imageDTOs to new package
